rename common to constants
